geo fence working

+++ app/src/main/java/com/omama/stationalarm/repository/StationRepository.kt
@@ -56,7 +56,7 @@ object StationRepository {
                 .addGeofencesForStation(
                     appContext,
                     activeStation.stationId,
-                    activeStation.alertDistanceKm
+                    activeStation.alertDistanceKm.toFloat()  // ← Add .toFloat()
                 )
         } catch (e: Exception) {
             Logger.log("ERROR", extra = "Failed to register geofence: ${e.message}")
